Add files via upload
So this was all a test file and an idea of what the program will be like for any developer.

The real work now starts with creating a database to allow users perform CRUD(Create. Read, Update and Delete) whatever they need in this inventory. Will decide what backend language I can use to connect this database to the frontend, but python's django or javascript might do the job if I don't continue challenging myself to keep up with Java.

If I do keep up, then I will use kotlin entirely(Though I suck at it cause I am currently learning it). The point of all this is to journal my progress as a developer student and trying too solve problems starting with agriculture. There's other projects but this is my first and anyone who helps me with it. I'd de grateful if I see this fruitful enough to sale..Thanks